+++ src/main/java/com/lowdragmc/mbd2/integration/kubejs/recipe/MBDRecipeSchema.java
@@ -29,6 +29,7 @@ import com.lowdragmc.mbd2.integration.pneumaticcraft.PNCHeatRecipeCapability;
 import com.lowdragmc.mbd2.integration.pneumaticcraft.PNCPressureAirRecipeCapability;
 import com.lowdragmc.mbd2.integration.pneumaticcraft.PressureAir;
 import com.lowdragmc.mbd2.integration.pneumaticcraft.trait.heat.PNCTemperatureCondition;
+import com.lowdragmc.mbd2.integration.pneumaticcraft.trait.pressure.PNCPressureCondition;
 import dev.latvian.mods.kubejs.fluid.FluidLike;
 import dev.latvian.mods.kubejs.fluid.FluidStackJS;
 import dev.latvian.mods.kubejs.fluid.InputFluid;
@@ -535,6 +536,14 @@ public interface MBDRecipeSchema {
                 throw new IllegalStateException("Try to add a temperature condition while the pneumatic is not loaded!");
             }
             addCondition(new PNCTemperatureCondition(minTemperature, maxTemperature));
+            return this;
+        }
+
+        public MBDRecipeJS pncPressureCondition(boolean isAir, float minTemperature, float maxTemperature) {
+            if (!MBD2.isPneumaticCraftLoaded()) {
+                throw new IllegalStateException("Try to add a pressure condition while the pneumatic is not loaded!");
+            }
+            addCondition(new PNCPressureCondition(isAir, minTemperature, maxTemperature));
             return this;
         }