+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## v1.0.29
 * Fixed Mek EMI support 
+* Fixed Photon node crash
+* Added an option for whether the block is `collisionShapeFullBlock`